"8:45am. Let me chill for a while. Any mail? A reject from UKG. Something from Indeed and SO. The Green Hills application has 72 applicants. Wow, that is a lot of competition. It says the employeer reviewed it. Unsubscribed from both.
8:55am. I want to chill, but I do not have anything I want to particularly do it with right now. Games are straight out, and I only have shitty novels waiting for me. Anime would take too long as well.

9am. Ok, let me read the Baki thread, this should be a non-made up translation.

9:10am. Let me start for the day.

I spent some time thinking about addressing and I think I understand it.

There are similarities between tabular RL and prop prog. Not in the algorithms themselves, but how the monadic implementation would work. There are also lessons I can take from tabular CFR when doing addressing.

Let me go through the intro to PPL book again. I want to make sure to absorb it properly. After that I am not sure.

Right now I have a complete grasp of Omega's semantics and will be able to use it for real tasks. I should also check out Turing which is another PPL in Julia.

9:15am. I am sure of it now - it is not deep learning, but prob prog that is the next evolutionary step after tabular RL.

9:20am. Let me start on the book.

138/218. Let me start from the middle.

9:25am. Hmmm, actually if it is poker, it would not be a bad idea to control the randomness the way a PPL does it instead of just sampling blindly.

...I messed up and added WebIO into the Omega library. Now it is precompiling everything again. Julia's compilation takes far too long even with all the staging.

9:35am. With prob prog, there is actually potential for me to combine it with tabular CFR. I never realized this before.

9:45am. Since I will be moving to Julia, it might be good to make time to read the dev docs.

https://docs.julialang.org/en/v1/devdocs/reflection/

Not today. Instead of a C backend, since the Cython one is so shitty, it might be a good idea to do a Julia one instead. A C backend would be helpful for an AI chip, but I do not anticipate getting it anytime soon. And it would be useless on its own as C has no libraries for me to use so I have no interest in working on it now.

Julia is different. It has a large ecosystem that a Spiral backend for it could take advantage of.

The game I made is extremely slow in Cython since object allocations are so expensive in it. But in Julia I would not be surprised for it to be on par with .NET. And unlike .NET, I could take advantage of Flux to do ML.

9:50am. Focus me. Let me read the book, and then I'll take a look at Zenna's examples. This time even if they are broken I have the knowledge I need to fix them.

One side quest I need to take on is to get familiar with how Julia does specialization. If I could do that, I'll be able to anticipate what kind of performance my programs will have. Right now I am blind as to how Julia optimizes.

10am. I am not really focusing on the book. Instead I am thinking how I could combine prob prog with different methods to train the poker agent on the side. But this is fine. Reading the book while giving my subconscious time to digest the recent gains in understanding and determine the next course of action is fine.

10:35am. 158/218. Let me focus on chapter six.

10:45am. 162/218.

> Support for forking can be implemented in a number of ways. In Chapter 4 we wrote evaluators that could be conditioned on a trace of random values to re-execute a program in a deterministic manner.

Actually, let me backtrack a bit to chapter 4. I want to make sure of how it does traces. As I said, I wasn't really paying attention to the book the first time around. Maybe if I had, maybe I would not have been so surprised by how prob spaces work in Omega.

11:35am. Ah, I see. If I am going to do nested optimization - Bayesian inference on top of sampling CFR for example, in order to make it deterministic what I could is make the random seed a top level variable.

Then that could be optimized as a part of the process. This is amazing.

That having said, NN ops can be nondeterministic which would wreck the idea, but tabular CFR would do fine. Seriously, what I am thinking here is wizardry. It would be unimaginable to the me of a few weeks ago.

12:45pm. 127/218. I do not entirely understand the acceptance criteria for MH, but nevermind that for now. Let me stop here for breakfast.

12:50pm. I'll watch some vids later to get a theoretical sense for it, but it does not matter at the moment. As I read the book, I am thinking how I could use the prob prog machinery for that agent. There are a lot of possibilities. Towards the end I'll even look at variational inference. I am going to cover all of it.

I am going to do it on the third run. The first one was a library building exercise, the second one with NNs was serious. They both failed.

If I cannot do it using probabilistic programming I'll accept my lot in life as a commoner. But until then I'll continue down this path of transcendence. As long as there is the slimmest chance to win it is my duty to pursue it."