"1:35pm. Done with breakfast. Let me read a chapter of Hero King, then the Satanophany raw. After that I'll resume the book. After that, I'll take a peek at that Turing PPL. I won't study the source like for Omega, I am just curious what state it is in.
1:50pm. One more chapter and then I'll start.

1:55pm. Let me resume.

2pm. This book really deserves praise on how clearly it is written. The algorithms in particular are easier to understand than anywhere else.

2:20pm. 133/218. Ok, I am not too interested in this. Let me skim these parts so I can get back to chapter 6.

2:55pm. 165/218. Here they talk about the addressing transformation. This is something I am particularly interested in.

3:05pm.

> In this rule, we begin by translating the expression e0, which returns a transformed function e 0 0 that now accepts an address argument. This address argument is updated to reflect that we are now nested one level deeper in the call stack. To do so, we assume a primitive (push-addr α c) which creates a new address by combining the current address α with some unique identifier c which is generated at translation time.

I am starting to understand why in Omega they took such a complex approach with the vectors for ids.

3:20pm.

> The result of this translation is that each sample and observe expression in a program will now have a unique address associated with it. These addresses are constructed dynamically at run time, but are well-defined in the sense that a sample or observe expression will have an address that is fully determined by its call stack. This means that this address scheme is valid for any HOPPL program, including programs that can instantiate an unbounded number of variables

This is more conservative than what I had in mind, but it is not wrong at all. I am overthinking things on my end for the sake of efficiency.

3:25pm. Ok, the addressing transformation is a piece of cake. It isn't really necessary to do it at every function call, the sample and observe expressions (when CPS'd) would be enough.

3:30pm. Let me take a break here.

3:40pm. https://turing.ml/stable/

1.4k stars, this is quite popular.

4:10pm. Let me resume. I'll go through the turing tutorials quickly.

4:15pm. I am still thinking about the addressing. Nothing is really satisfying to me.

Actually, instead of all this complicated stuff, shouldn't it be possible to just fetch the address pointer at the current address and use that? That would be ideal for something like Omega. No need to give an unique id to every statement.

Ah, no that would not work with multicore. Then the current source location would not be a bad alternative. Something like a string that could then be compared using ref equality. That would work great. Much better than these global counters.

Ok, focus me on the Turing tutorial. What did the Omega docs have to say about it?

> Turing: Omega does not yet suport likelihood based inference, although it is planned, so use Turing if that is the case. Turing currently has more infernece procedures implemented. Omega is more flexible, likely faster, and supports causal and distributional inference.

Is it really faster? I'd guess that Omega is far from optimal, though I know nothing about Turing yet.

> In Omega you condition on predicates
> Omega supports likelihood free inference
> Omega supports causal inference
> Omega is fast (and type stable)

Is it really type stable? I am going to have to familiarize myself with this aspect of Julia. But a lot of the code examples would not be possible to type in first order logic, not unless Julia pevals very aggressively.

Let me add Turing to the package repo. It will take forever so I might as well take a head start.

4:50pm. Lunch time.

5:50pm. Let me do the chores here.

6pm. Ok, I've skimmed the Turing tutorial. It is boring statistics stuff, not nested optimization which I'd need for a poker agent.

What is next? I'd like to at least take a peek at the Turing source. Let me clone it.

6:15pm.

```jl
export  @model,
```

Damn it, where is this macro?

6:20pm. No I can't find it. Turing's source code is nothing like Omega. It is not even large.

I'll study it tomorrow. I want to see what magic does it do.

https://github.com/TuringLang/Turing.jl/blob/3b46cd0a469891009fc9b61ffb9b66cbef779cff/docs/src/for-developers/compiler.md

> In this section, the current design of Turing's model "compiler" is described which enables Turing to perform various types of Bayesian inference without changing the model definition. The "compiler" is essentially just a macro that rewrites the user's model definition to a function that generates a Model struct that Julia's dispatch can operate on and that Julia's compiler can successfully do type inference on for efficient machine code generation.

Yeah, where is it?

https://github.com/TuringLang/DynamicPPL.jl

6:25pm. I'll have to study DynamicPPL if I want to understand Turing. Turing itself is just the frontend for this.

6:30pm. This is a good spot to stop for the day and get some game time in. Given the direction I am going in, studying Turing would not be a bad idea at all.

Besides this, what I wanted to do was finally move to Zenna's examples. I still haven't really found the perfect scheme for addressing in PPLs. I said about getting the source location, but that would not work due to type reflection capabilities of Julia. To make it work correctly I'd need the source of the monomorphic compiled code.

This is hard.

6:35pm. Well, it is fine. I'll figure it out eventually. Let me do it all properly here. The next things on my list are the Turing source and Zenna's examples. With effort I will get through them to where I want to get to.

I cannot be held back by this. PPLs do have the potential to surmount the poker obstacle, but I need to cultivate my understanding of them first. This is what I am supposed to be doing.

I was too naive to think I'll be able to get anywhere in AI without mastering probabilistic reasoning first. If having to learn PPLs is my punishment I will accept it wholeheartedly."