"8:55am. I am really into Ender Lilies, so much that I've been slipping into a habit of going to bed later just to get more game time in. Yesterday's gaming session was immensely satisfying. This time the Youtube sidebar did a really good thing by recommending me this game's OST.
On a side note, as annoying as Anti was, he is probably right that Hollow Knight is great. I should give it a try at some point in the future.

9:05am. Any mail? No.

9:35am. Let me start. What should I do first?

Let me check out if Julia has effect handlers.

https://github.com/MikeInnes/Effects.jl
> [Caveat Emptor: Julia's compiler is not designed to handle this kind of code and reserves the right to complain / be slow. Also, while everything theoretically composes nicely, this is not thoroughly tested.]

http://pyro.ai/examples/effect_handlers.html

Pyro has effect handlers? That is interesting.

One choice that I have is to stick with Python and check out Pyro.

https://discourse.julialang.org/t/minijyro-a-toy-pyro-like-ppl/37356/2
> Turing contexts and Pyro effect handlers seem almost identical. Both target modifying the behavior of a probabilistic program without manually modifying it. Turing contexts have zero run-time overhead though thanks to the beautiful Julia compiler.

> The zero run-time overhead can be achieved because Julia compiles a different method for each context type. There is nothing Turing-specific here. We just have different types for different contexts and use multiple dispatch to define the specialized behavior of each context. Then Julia does the rest.

http://pyro.ai/examples/effect_handlers.html
Poutine: A Guide to Programming with Effect Handlers in Pyro

9:50am. I am familiar with DL, but PPLs are new to me.

http://pyro.ai/examples/intro_part_i.html

It seems that Pyro also uses explicit addresses which is fine for Python. I guess I'll study these tutorials.

9:55am. Yeah, I can't just ignore what is possible on the Python side of things. Whether I want to use Python or Julia depends whether or not I want to stick with NNs. If I want to try a no-NN run I am going to need Julia's speed benefits.

10:25am. I am having a crystallizing experience as I think of the implications of using PPLs to improve tabular CFR.

10:45am.

> Pyro is built to enable stochastic variational inference, a powerful and widely applicable class of variational inference algorithms with three key characteristics:

11:05am. It is all coming to me. Let me take a short break here.

11:25am. Yeah, it is coming to me. When I first saw CFR I was really surprised. Unlike RL, it did some kind of Bayesian thing with the averaging of categorical policies. The reason this was so mindblowing is because it extended vanilla tabular RL which I could not even imagine how to do at the time.

11:30am. What would be the next step past that? I thought that NNs were it. But I was wrong. CFR is is really a probabilistic program with just a single categorical variable followed by an observation of the reward.

No, not CFR itself - a single node is what I am refering to.

Then the whole of CFR would be a nested Bayesian optimization process. Vanilla sampling CFR cannot account for variance reduction. Using the game state as node addresses would just make it MC.

It is possible to go further. Replace the uniform policy with a richer probabilistic program, and what tabular RL gives us is a way of doing nested Bayesian inference. Nothing else makes sense for propagating credit across arbitrary program nodes.

Nested Bayesian inference is just asynchronous inference. As long as there is a controller to distribute the work to different nodes it will work. The only real requirement is that the traces be non-recursive.

This requirement is something that deep RL breaks absolutely. It is probably the main reason why the optimization is so harsh there.

Yeah, I was aware that I was breaking the non-recursive trace requirement by moving to deep RL, but I did not know how to deal with it.

12:10pm. Let me have breakfast here. I spent all this time in thought instead of reading the Pyro docs, but who cares about that.

I feel like I have it. I am on the cusp of breaking into 3/5 in ML. In order to actually do it I will have to execute and practice my plan, but it will be doable with effort. Without a doubt, I'll be able to make that poker agent if I follow this path.

12:20pm. I think that seeing Gen and now Pyro use addresses explicitly must have jogged my mind. My first instict is see this as a design error, but in fact explicit addressing is what would allow me to bridge tabular RL and PProg.

I came up with the elegant trace scheme yesterday, but there is no rule that it has to be done that way. In fact this way is easier as I can do `replace` and causal inference much more readily.

12:25pm. Addressing is at root of everything in PProg. It seems like a side thing, but it is really a front and center concern.

12:30pm. I've changed my mind on Gen. It is not a bad choice to make addressing explicit even if adds verbiage. It would be easy to make a check to make sure that the same var is not being sampled from twice.

The issue of addressing is something I should be flexible on. In fact, the reason why Bayesian math is so confusing is because no address is associated with it. Neither is CPSing.

The lens of probabilistic programming is the right way to approach Bayesian reasoning.

12:40pm. Let me have breakfast here instead of just ranting.

This is so great. A month ago when I threw in the towel I was broken, but now I see hope again. I am really lucky after all."