"1:35pm. Done with breakfast. Let me chill a bit. Today seems to be a thinking day. Let me finish up here, then I'll go through the intro PyTorch tutorial that I've been staring blankly for a few hours. After that I'll look into nested inference papers.
2:35pm. Let me resume. I've been in my own world today. I am not even distracted by fiction, just my own thoughts.

Today's ideas are huge. Cracking nested inference is what I need to make the agent viable. And now the path leading to the solution has been discovered. I now have a concrete goal instead of doing random prob prog work in hopes of getting inspiration.

My starting point - tabular CFR, was not wrong. Where I went wrong is betting it all on deep learning. That was not the right path to go down. Instead I should have used CFR as a model to generalize probabilistic programming. I'll have to look around to figure out whether I can adapt any of the existing libraries for nested inference. But if not, I'll just do my own thing. It is not a problem.

http://pyro.ai/examples/intro_part_ii.html#Flexible-Approximate-Inference-With-Guide-Functions

Focus me. Let me just read this.

2:45pm.

> What we can do instead is use the top-level function pyro.param to specify a family of guides indexed by named parameters, and search for the member of that family that is the best approximation according to some loss function. This approach to approximate posterior inference is called variational inference.

If I want to learn about VI, I should study Pyro. It is made for that purpose.

I am not going to be able to get away with using something like SSMH. Because I no longer have local rewards, I'll have to use some kind of gradient method in order to adjust the proposal distro.

2:55pm. In addition to nested inference, I should look up Hinton's dynamic routing.

3:20pm. Ah, let me take a break. I can't focus on the PyTorch tutorial even a little.

I suppose days like these are fine on occassion. This warants a singificant change of plans. Forget Z's examples and Omega.

I am going to study and play around with variational inference in depth. I'll get some experience actually using the existing libraries.

3:55pm. Let me resume. I can't believe it is already almost 4pm.

Pyro and Gen tutorials should be my main focus for the next few days. If in the future I get paid part-time work on PPLs that is fine, but if not that is fine too. Now that I have my current ideas, the quest is alive again. I know that all I have to do is master PPLs and I will have my elite agent. This time I will succeed because not only will I have all the strengths of tabular CFR, I will be able to use PPLs to build in biases that will vastly increase the efficiency of training.

I just have to put in the work, and I will accomplish my long cherished dream.

These useless programming skills that I've been so painstakingly refining for years and years will become actual power.

4:30pm. http://pyro.ai/examples/svi_part_ii.html

4:45pm. http://pyro.ai/examples/svi_part_ii.html#Subsampling-when-there-are-only-local-random-variables

There is a lot of that I do not understand about VI.

http://pyro.ai/examples/svi_part_iii.html#Variance-or-Why-I-Wish-I-Was-Doing-MLE-Deep-Learning
> Variance or Why I Wish I Was Doing MLE Deep Learning

Lol.

5:20pm. http://pyro.ai/examples/svi_part_iv.html

Let me skim this and then I'll have lunch and call it a day. I can feel the hands of fate pushing at my back. It is time to get serious about the quest again. I am going to figure out the right way to make that agent. I will find a way to do it that would be good even without AI chips. I will be able to make huge strides with probabilistic programming. As long as I keep cultivating this skill, I will be able to break through to 3/5 which will be enough to make the agent, and then 4/5 which will make me a force to be reckoned with.

There is a huge amount of work waiting for me, but I will surmount it.

5:35pm. http://pyro.ai/examples/bayesian_regression.html

Let me stop the Pyro tut here.

https://arxiv.org/abs/1710.09829
Dynamic Routing Between Capsules

> A capsule is a group of neurons whose activity vector represents the instantiation parameters of a specific type of entity such as an object or an object part. We use the length of the activity vector to represent the probability that the entity exists and its orientation to represent the instantiation parameters. Active capsules at one level make predictions, via transformation matrices, for the instantiation parameters of higher-level capsules. When multiple predictions agree, a higher level capsule becomes active. We show that a discrimininatively trained, multi-layer capsule system achieves state-of-the-art performance on MNIST and is considerably better than a convolutional net at recognizing highly overlapping digits. To achieve these results we use an iterative routing-by-agreement mechanism: A lower-level capsule prefers to send its output to higher level capsules whose activity vectors have a big scalar product with the prediction coming from the lower-level capsule.

I'll take a look at this tomorrow."