"9:15am. I went to bed early and had time to think. Forget LNNs. I was euphoric yesterday about real valued logic being connected to NNs. It is true that this would make things explainable on a small scale. I could make a hand crafted player with a few dozen parameters. But what about the other million? Would I be able to interpret that as well? The autotrained stuff in the middle layers?
Of course not. And furthermore, I do not have any better plans than providing a few extra features. What am I going to with OR and IMPLIES?

Ultimately, I am just hiding from the fact that I've failed and looking for outs.

9:25am. If I went the LNN route it would still be yet another hacky method for me to try. It would be another few months down the drain. Gary Marcus said that a good check is: Can I try it out myself? Is it roboust?

The story is pretty good, but it is a red flag that there is still no code for this out there.

If my goal was to do first order logic via neural nets then there would be no contest - I would have to give this a try. But it is not. I want to make a poker agent.

9:30am. Ultimately what I really have to do is change my target from hacks, to modeling the other player.

And I should not forget - the simple RL approach failed and there is not much I can do to turn this ship around. I started thinking I could win, but that is not true.

Forget winning. I am in a spiral into the Abyss. The most I can do is learn to control randomness and master probabilistic programming. This is the true framework. Every time I stray the course I end up being punched in the face. You'd think I'd learn by now to ignore things that lead me away from rational reasoning in machines.

I've been convinced of the benefits of Bayesian reasoning years ago when I went through the Prob Mods book. All the other paths lead to even surer ruin.

What I should do now is try the approach of modeling the other guy. Make probabilistic programs. Try to extract efficiencies that way.

If I went down the LNN route, could I make simulators and things of that sort? No, I'd be tied up in expressing things in first order logic. Whereas probabilistic programming make it possible to bring all my general programming technique to bear upon the problem.

9:35am. Forget hacking addressing and nested inference for the time being. Forget categorical RL and hacks of that sort to make the net work better.

If I am going to get significant efficiency gains, it won't be by scouring the literature. I need 1,000x and more in efficiency. I won't get that by hacking the way the rewards are distributed, but by fundamentally changing the model in ways that regular deep learning won't allow me to.

9:40am. I need to do the most straightforward thing here. I've tried everything except the sane thing.

I need to supress my skepticism and stop looking for outs. Instead just get a PPL, and do what is needed using it. Forget trying to hack things.

Don't target the rewards. Target the other guy's hand. This is the thing I haven't tried and LNNs won't allow me it.

I need the full power of PPLs to tackle this problem.

9:50am. I can't build things into the NN directly which the LNN approach would require me to take.

Instead what I can do is build simulators.

I can retrace my steps and take control of the randomness that I had before and optimize that together. Remember how in the old thing I had to sample actions because they were too many of them to express them through a single catty distribution? Or the other sources of randomness like hand selection.

I need to implement the whole system through the lns of PPLs. That will allow me to do inference backwards through the simulator. This is my only hope for making a significant improvement.

When I was doing DL, my focus was just on shoving data into it. But with PPLs I can take a step back and consider the whole system as a single entity.

This is my only hope for winning. LNNs would not allow me to take this approach. They wouldn't be more efficient than just feeding a regular net directly.

9:55am. It is not like I am really trying to do anything fundamentally new.

Rather what I need to do is take my old programs and draw out their possibilities in ways that deep learning does not allow me to. I've gotten lead astray by my doubts.

One thing I should hope for is that predicting the opponent's hand and using that to calculate EV instead of just feeding the rewards will be a better learning objective.

Yeah, I do not think I was wrong for not trying out GANs during the last run. But the approach might become viable if I restructure it fundamentally.

It is true that there are things I have not tried, but none of those things are in the standard deep learning framework.

10am. Ok, now let me finish reading the Urasekai chapter (I've been stuck on it for 40m) and I will start.

I'll resume the Pyro tutorials. This is what I need to focus on. I need to get back to studying PPLs.

A few weeks ago I was applying to jobs. I can still do that in the future. But for the next few months or longer, let me just focus on this. Forget deep learning and master probabilistic programming. That will finally put my ML skills on a firm foundation.

Probabilistic programming is the last thing I am going to try. There is nothing beyond it.

But before I've fully exhausted those possibilities, I should believe in it and strive to go forward. The thread is not broken and I am yet to pay my respects to rational reasoning.

If it is a year or two it does not matter. There will be plenty of time to boring paid work in the future if I fail.

10:10am. http://pyro.ai/examples/svi_part_iv.html

I finished this last time. Let me take a look at it again and then I'll go through the rest of examples.

...Yeah, forget about ever feeding the rewards. How about I make that a design constraint? Instead of doing the RL thing, I should aim to build a simulator and calculate the EV of various actions based on the estimate of what the opponent is holding.

I should model what the opponent is holding, how aggressive he is and how likely he is to fold and then use that to determine the course of action. I won't ever try just feeding the rewards to an RL system ever again.

This is justice. It is the right thing to do.

A few years to do this should be enough. I'll leave the speedruns for some other life.

10:15am. http://pyro.ai/examples/bayesian_regression.html

Ok, let me read this. I should focus on going through the Pyro and Gen tutorials for now. After all that is done I'll ease myself into doing programming work.

10:30am. Let me take a break, I do not feel like reading this right now. I've been burning my brain yesterday and this morning trying to decide on a course of action.

Forget Pyro. If I am going to be doing simulators, no way will Python be fast enough. I have to do it in Julia.

I need to reafirm that goal. I am bored to death of reading things by this point. I need to get back to Julia and do some actual programming in it.

I should take a look at the rest of Z's examples and then play with the Gen and Turing libraries. I should look at how RL is done in Julia.

10:35am. Right now let me take a break. Since I skipped dinner yesterday I might as well have early breakfast here.

...Hmmm, now that I think about it, the most basic essence of proof is that for a proposed plan or sequence of actions that one be able to imagine a course of action that would be beneficial. Being able to predict the future is a proof technique.

Current RL systems and what I was doing was missing that. Whenever one is wrong, one gets punished. The beatings will continue until my policy improves. I did gain the ability to write games and simulators from my long days of practice. I should put it to good use."