"5:10pm. If it is just doing VNs, I am sure that Unity is overkill, but it is the first thing that pops to mind. I should focus on writing out 100 different scenes first and then consider how to fit them.
I should go down this path.

I feel bad that I won't be able to make the money needed to cryofreeze mom in order to save her from cancer, but that is how the dice roll. Life is pretty brutal. She is unlucky to have contacted it at this time. I treasure the time she is still here.

I might be hard working, but I am also filled with evil.

I am going to change. Just as I've changed my personality in 2014, and then revisited programming in 2015 and kept it up till today, writing is also something that has taken its hold for me. Everything that I like should be brought together.

If my luck and ability is not good enough to make it this way then so be it. I could always get some lethal disease myself, or might heart might stop out of the blue. Or I could fall behind in the race and the winner might just decide that other people should not be given a chance to self improve ever.

5:20pm. The way hiring works is like some dystopian comedy. Just why is every HR person I've gotten in contact so far some woman? It is like the very important job of evaluating potential candidates all went to people least qualified to judge them. Having Spiral on my resume counts for shit.

And my ability to bullshit is not that good since I do not have any experience at companies. I'd sure like to know how people get recruiters to hound them constantly.

I am not going to try this anymore. I am just not good at playing this game.

5:25pm. I should gather my will and focus entirely on my own work, plus the threads I am interested in. I am going to get rid of that freelancer resume. If I find something of interest on ai-jobs or compiler related work I'll use the 3 page one. I'll ignore that guy who offered me advice.

I feel that most of that advice is some kind of demoralization work. Or for all I know the people who are offering it might be morons.

There should be a way to just be honest and say that I've never worked at companies before, but I am looking forward to trying it. And that I have the relevant experience for the role anyway. After the week is done, I'll try sending the 3 page resume at Intel again.

5:30pm. A lie is only good if you know how to deliver it properly. If I was a recruiter and I read that buzzword resume, I'd be insulted. For the backend positions they would not care about the resume and like Theorem just send me to Devskiller or HackerRank anyway.

5:40pm. Let me read Gunnm. I do not feel like writing today. I barely even feel like playing. This feeling of unease is quite something. 100 scenes. Let me go 100 scener or so, and then I will look into GANs for the first time. That will be my triumphant return to ML.

I need something I can show to people. If I could have done that poker tutorial, my position would have improved significantly.

I must not make the same mistake again.

5:55pm. Let me get lunch. I'll close here for the day. I am almost done with Hollow Knight, only the extra endings and the add-on content remains. The game deserves the praise it gets. It has charm in spades."