"9:35am. Sigh, I need to temper my obsession a bit.
Right now, my biggest source of anxiety is the CPS transform part of a PPL. Doing that in Julia would kill performance, so the best alternative would be to make use of coroutines. But I do not know how to do the later or the former in Julia for that matter, so if a task got sprung on me, I'd be missing a vital piece.

Well, since he gave all that advanced stuff to me, I do not think I'll get work right away. He'll most likely give me 2 weeks to get the ball rolling, and given how busy he was last month, probably a full month or even longer. I need to make use of that.

This whole issue is kind of annoying too as only SMC and similar particle methods really need this capability. If I was doing my own PPL I'd be focusing on VI and would not even care about it, but since I want to do a pro job, I need complete coverage.

Any mail? My god, he is trying to reduce the quoted rate by 10x, the nerve! Well, I knew it was too good to be true, so I was not unprepared for him changing his mind, but this is too low.

10:20am. The rate he is offering now is 2.5k per month. I need to bring this to at least 6k, he is looking down on me too much. Maybe I'll be writing Simulacrum after all. It is a bit risky to bargain since he might pull back, but 2.5k is not attractive for me as a starter. If it is that much I'd rather just do my own thing.

From here on out, I need to remember to check my mail at 10pm mark since this is the time he seems to be active when replying.

10:25am. Sigh, this is such a roller coaster ride.

Also my assumptions seem to be wrong and he wants to start right away. Well, let me focus on the Cassette tutorials today.

10:30am. Let me try to leave this out of mind. I should get familiar with the basics of the advanced stuff.

11:05am. Ok, Cassette is a bit of a troublesome case, because to make use of it not only would I need to understand Julia internals, but the way it works as well.

Whereas with a macro based approach, I'd only need to know the first and there isn't so much black box stuff. I am partial to the Soss approach here. The Turing approach on the other hand will secure their implementors a job for life.

> Information on Julia's IR - https://docs.julialang.org/en/v1/devdocs/ssair/
> Nonstandard / contextual execution - https://github.com/JuliaLabs/Cassette.jl
> Tutorial on Type inference through abstract interpretation -  https://aviatesk.github.io/posts/data-flow-problem/
> Partial Evaluation and Abstract Interpretation (Mjolnir) - https://mikeinnes.github.io/2020/07/29/mjolnir.html

Ok, that is first 2 down.

I've looked at the last two yesterday. Let me just do a quick skim. Now that Z told me he wants to start right away, I am feeling pressed for time.

I need to look into the latest Omega, and after that study the coroutine mechanisms.

11:20am. Focus me.

https://github.com/zenna/Omega.jl/tree/lang

Let me clone this again. I need to see how far he got.

Focus me.

11:40am. It seems it has changed a fair bit. The 5 days I spent studying the original will have to be redone. I'll have to spend a few days getting familiar with it, but as far as I can tell the approach itself did not change.

Let me leave that aside. The library itself is not the source of my anxiety. Right now, the CPS transform is the most troublesome. I need to get the benefits of it using coroutines in order to be able to do proper `observe` and `sample` statements like in the `Intro to PPL` book.

Things like propagating likelihoods and CPSing is something I've done for just the poker game already. But Spiral is the only language in the world that can do the later sort of thing efficiently. I know that Julia is bad with both monads and effect handlers.

11:50am. I need to figure out how to do SMC style algorithms in Julia using yieldto. Though I am an expert in high level concurrency patterns thanks to all my work on Spiral, I am not familiar with the low level aspects. Neither am I familiar with coroutines. Julia supposedly has coroutines, but they aren't anything what they seemed to be in Scheme. In Scheme they are like resumable exceptions and can be used to implement effect handlers.

But it is not like I am familiar with effect handlers either, I just know what they are.

Supposedly Turing uses coroutines to implement its functionality. I could use it as a reference if only it weren't so opaque.

https://docs.julialang.org/en/v1/base/parallel/#Scheduling

I wish there were some examples of how this works. Well, I have an idea how to experiment with it.

Let me take a break here."