"1:45pm. Let me resume. Let me watch some vids on tracing just for fun.
https://www.youtube.com/watch?v=AvFmt38Y2jg
Should artists TRACE? • Tracing, studying and referencing (examples / speed paint)

Once I get serious, instead of semi-plagiarizing other artists, I should just trace Clip studio models. That video I saw yesterday was amazing, but there is no way I would ever be able to approach his level of skill in the near term.

2:10pm. This video is so annoying, let me stop at 2/3rds.

https://youtu.be/k0ufz75UvHs
Iterative Drawing - The Fastest Way to Improve

Let me take a look at this.

2:20pm. Forget those ranty videos. They are wastes of time.

https://www.youtube.com/watch?v=wAOldLWIDSM
Quickly Draw Heads with the Loomis Method - Part 1

Let me take a look at this.

2:30pm. Ok, forget these tutorials. Let me just install Adobe Illustrator and start going over the tutorials for it.

https://youtu.be/_raUbrGdImk
Is It WRONG to TRACE Art? - Digital Artist Doubts

Let me watch this guy while the thing installs.

2:35pm. This is a pretty good video. He says that tracing a good way to train hand eye coordination.

2:50pm. https://www.youtube.com/watch?v=N-buLx_tV6U
THE BEGINNER'S GUIDE TO DRAWING

Let me take a short break, then I'll watch this, then I'll start going through the tutorials for real.

3:20pm. Let me watch this guy.

I the last 6.5 years, I think my failed RL attempts are what hurt me the most. I put the most hope there, and I got the least return.

I think to go forward in life, it is important to just keep improving. And RL did not really give me that.

I need to get back to the basics.

NP Hardness of learning is the real problem. NNs offer a way to do randomized search in a approximately memoizing fashion and constrain the search space that way.

In this time we are too used of thinking about passing a static dataset to them, because the current hardware is very hard to use in an online fashion, but that will change. Then the true arc towards the Singularity will start.

I really did not want to start applying to jobs. That is not what would fill the hole in my heart. Rather what I need to do is find a new path.

I won't aim to conquer the world, instead I just have to find my niche.

I need to get back to the basics. My great belief got me this far, and I need to find what the next step is. Maybe doing art for a few weeks will give me some inspiration.

> Going from zero to pro takes only 5 years of investment and guided studies.

I want to do this in less than a month. I can only hope that my programming ability gives me a boost. I really can't practice for 5 years before doing the art for Simulacrum.

https://youtu.be/N-buLx_tV6U

This is a pretty good guide. That information that is not needed is discarded by the brain is completely true.

3:50pm. Let me start. It is time to get grinding.

https://helpx.adobe.com/illustrator/tutorials.html

I am going to keep applying to jobs, but not with the intention of getting them. Like that HN position that was open, I'll want to send my 3 page resume and talk about my vision of ML. That is what I should be focusing on.

Yeah, I am not sure I really wanted to work any of these jobs. Rather I just want to keep improving. Because I can't improve in RL, that made me lose my sense of direction. I still haven't found it again, but I will.

I need to get back to basics.

Making games is easier than solving them.

I've made a mockery of that, and tried jumping limits. I should have been wiser.

I need to put things in perspective.

I had a conjecture and it did not turn out to be true. This is valuable experience that the me of 6.5 years ago did not have. I did learn quite a bit. Especially these latest insights on how to do counterfactual MCTS. That is not something a beginner could come up with. I've taken the algorithm and generalized it to tree search and related it to probabilistic reasoning making it act as support.

In the future this will be a small, but important part of a more general way of programming.

But now, why don't I learn to do games starting with LNs. I want to promote the path of evil. Attacking the prevailing culture is always a good way to garner attention, as long it is done in a sensible way. Last time doing the story set me on this path, but this time it will be different.

4:05pm. Let me run the illustrator for the first time.

4:10pm. Yeah, it works, no problem.

4:15pm. Focus me, let me watch the videos. Real tutorials, not that Youtube tablet stuff.

4:30pm. https://www.reddit.com/r/MachineLearning/comments/pzqgvb/d_thoughts_on_neural_nets_simulation_based/

I went to the ML sub and found this. Let me finish the zooming tutorial and then I'll listen to this.

4:50pm. Let me watch the interview with Lex on the part that was posted.

4:55pm. Ah, no nevermind. He is arguing against simulation. Let me go to the next tutorial. I need to keep going at it. Master the Adobe Illustrator. That should be my goal.

5pm. This is so boring, I can barely stay awake.

5:10pm. Focus me.

https://helpx.adobe.com/illustrator/how-to/shapes-basics.html

Let me go through this. Once I go through these tutorials, I'll be able to start my tracing adventure. First I learn to trace, then I learn to adjust the trace, and pretty soon I'll be a serious artist.

5:30pm. Next! Combine shapes.

5:40pm. Next is tracing. I think after that I'll call it a day.

6pm. https://helpx.adobe.com/illustrator/how-to/edit-artwork-basics.html

Let me stop here. It is time I get lunch and watch Son of Ogre.

6:05pm. Art design much like UI work and concurrency previously, art design is really one of the aspects in which I've felt inferior in ways that stiffled my development. If I can master this aspect along with sound it will go a long way in enabling me to live with my own power."